#8 finished

+++ src/controllers/videoController.js
@@ -32,19 +32,28 @@ export const watch = async(req, res) => {
 
 export const getEdit = async(req, res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
 
     if(!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."});
     }
+    console.log(video.owner, _id);
+    if(String(video.owner) !== String(_id)){
+        return res.status(404).redirect("/")
+    }
     return res.render("edit", {pageTitle: `Edit ${video.title}`, video});
 };
 export const postEdit = async(req, res) => {
+    const {user: {_id}} = req.session;
     const { id } = req.params;
     const { title, description, hashtags} = req.body;
     const video = await Video.exists({_id: id});
     if(!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(404).redirect("/")
     }
     await Video.findByIdAndUpdate(id, { 
         title,
@@ -66,7 +75,7 @@ export const postUpload = async (req, res) => {
     const { title, description, hashtags } = req.body;
     console.log("eevee", path);
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl: path,
@@ -75,6 +84,9 @@ export const postUpload = async (req, res) => {
             createdAt: Date.now(),
             hashtags: Video.formatHashtags(hashtags),
            });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save(); 
         return res.redirect("/");
     } catch(error){
         return res.status(400).render("upload", {
@@ -87,8 +99,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async(req, res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(404).redirect("/")
+    }
     await Video.findByIdAndDelete(id);
         //de;ete video
+    
     return res.redirect("/")
 
 };
